requirements.txt
pymavlink==2.4.42
PyYAML==6.0.1